make recoverable public and add migration guide.
