+++ app/src/main/kotlin/pro/respawn/flowmvi/sample/view/LambdaViewModel.kt
@@ -9,7 +9,7 @@ import pro.respawn.flowmvi.android.plugins.androidLoggingPlugin
 import pro.respawn.flowmvi.android.plugins.parcelizeState
 import pro.respawn.flowmvi.api.Container
 import pro.respawn.flowmvi.api.PipelineContext
-import pro.respawn.flowmvi.api.lazyStore
+import pro.respawn.flowmvi.dsl.lazyStore
 import pro.respawn.flowmvi.dsl.reduceLambdas
 import pro.respawn.flowmvi.dsl.send
 import pro.respawn.flowmvi.dsl.updateState
