+++ core/src/commonMain/kotlin/pro/respawn/flowmvi/modules/Recoverable.kt
@@ -3,25 +3,14 @@ package pro.respawn.flowmvi.modules
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.launch
+import pro.respawn.flowmvi.api.DelicateStoreApi
 import pro.respawn.flowmvi.api.MVIAction
 import pro.respawn.flowmvi.api.MVIIntent
 import pro.respawn.flowmvi.api.MVIState
 import pro.respawn.flowmvi.api.PipelineContext
-import kotlin.coroutines.CoroutineContext
+import pro.respawn.flowmvi.api.Recoverable
 
-/**
- * An entity that can [recover] from exceptions happening during its lifecycle. Most often, a [Store]
- */
-@Suppress("FUN_INTERFACE_WITH_SUSPEND_FUNCTION") // https://youtrack.jetbrains.com/issue/KTIJ-7642
-internal fun interface Recoverable<S : MVIState, I : MVIIntent, A : MVIAction> : CoroutineContext.Element {
-
-    override val key: CoroutineContext.Key<*> get() = Recoverable
-
-    suspend fun PipelineContext<S, I, A>.recover(e: Exception)
-
-    companion object : CoroutineContext.Key<Recoverable<*, *, *>>
-}
-
+@OptIn(DelicateStoreApi::class)
 @Suppress("FunctionName")
 internal fun <S : MVIState, I : MVIIntent, A : MVIAction> Recoverable<S, I, A>.PipelineExceptionHandler(
     pipeline: PipelineContext<S, I, A>,
@@ -35,6 +24,7 @@ internal fun <S : MVIState, I : MVIIntent, A : MVIAction> Recoverable<S, I, A>.P
     }
 }
 
+@DelicateStoreApi
 internal suspend inline fun <S : MVIState, I : MVIIntent, A : MVIAction> Recoverable<S, I, A>.catch(
     ctx: PipelineContext<S, I, A>,
     block: () -> Unit
